Move purge and fix methods
